requirements.txt
openai>=1.30.0
edge-tts>=6.1.9
python-dotenv>=1.0.0
requests>=2.31.0
python-telegram-bot>=21.0.0
Pillow>=10.0.0
